Updated database query and added caching technology
Updated the database query such that it only considers posts made within the past x amount of days for the trending topics algorithm
Added caching for trending topics such that when the top keywords for a new collection date are generated, it stores the result in the database called 'caching' under the collection 'trending topics'.  When that date is accessed again, instead of generating them, it just fetches the corresponding value from the database.