Added caching technology for trending topics
Added caching for trending topics such that when the top keywords for a new collection date are generated, it stores the result in the database called 'caching' under the collection 'trending topics'.  When that date is accessed again, instead of generating them, it just fetches the corresponding value from the database.